Rework design system page + separate stories
