+++ packages/front-end/ui/Stepper.stories.tsx
@@ -1,0 +1,22 @@
+import { useState } from "react";
+import { Flex } from "@radix-ui/themes";
+import Stepper from "@/components/Stepper/Stepper";
+
+export default function StepperStories() {
+  const [stepperStep, setStepperStep] = useState(0);
+
+  return (
+    <Flex gap="3">
+      <Stepper
+        step={stepperStep}
+        setStep={setStepperStep}
+        setError={() => {}}
+        steps={[
+          { label: "Step 1", enabled: true },
+          { label: "Step 2", enabled: true },
+          { label: "Step 3", enabled: true },
+        ]}
+      />
+    </Flex>
+  );
+}
